Prevent arbitrary HTTP fetches for map tile proxy mode
Certain clients enforce internet quotas on their employees, and so we
act as a proxy for those users, fetching map tiles on their behalf,
from third party map providers.

This is too lenient, so this commit brings with it the restriction that
only whitelisted hosts can be contacted.

Ticket TI-3784