Remove URL rewriting in HTTP headers
As far as I can tell, this old code was only there for the
sake of Yellowfin. It was horribly flawed anyway - doing a
simple string find & replace on the entire URL line, which
in fact it should have been doing that only on the hostname
portion. However, if we can get rid of this completely, then
we should.

If this commit doesn't cause any issues for a month or so,
then we can get rid of the old commented-out code.