feat: Adds forwardHTTPSSE (#16)
Add the ability to route sse (server sent events). This primarily requires the ResponseWriter to support the http.Flusher interface which apaches logger wrapper did not allow. A change was made to make the apache record public and using this the ResponseWriter was exposed to allow for the flusher interface.

Further a check is done to router Content-Type text/event-stream to the correct handler.

Also, since our router does not speak https to the BE services, the connection is upgraded using h2c to allow the whole process to work.

Refs: ASG-4475, CT-1121